pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ii-agent-telegram-bridge"
version = "0.1.0"
description = "A bridge server for ii-agent to use a Telegram bot"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Joe Heffer", email = "jheffer@gmail.com" }]
dependencies = ["pyTelegramBotAPI", "requests", "python-dotenv"]

[project.optional-dependencies]
dev = ["ruff"]
test = ["pytest", "pytest-cov"]

[project.scripts]
telegram-bridge = "telegram_bridge.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["telegram_bridge"]
